Feature/sprint6 t7 t9 (#359)
* Raw client, post_bulk method in automatical endpoint generation

* Removed unused imports

* Test fix integration tests, added more tests

* Try fix tests

* Try fix tests

* Bump version of client library

* Bump version of core library

* Modifications to codecov.yml

---------
