+++ libs/client/cornflow_client/__init__.py
@@ -1,4 +1,5 @@
-from .cornflow_client import CornFlow, group_variables_by_name, CornFlowApiError
+from .cornflow_client import CornFlow
+from .raw_cornflow_client import group_variables_by_name, CornFlowApiError
 from cornflow_client.schema.manager import SchemaManager
 from pytups import TupList, SuperDict, OrderSet
 from cornflow_client.core import (
