+++ libs/client/cornflow_client/airflow/dag_utilities.py
@@ -154,6 +154,7 @@ def cf_solve(fun, dag_name, secrets, **kwargs):
     :param kwargs: other kwargs passed to the dag task.
     :return:
     """
+    config = dict()
     try:
         client = connect_to_cornflow(secrets)
         exec_id = kwargs["dag_run"].conf["exec_id"]
