+++ libs/client/cornflow_client/tests/integration/test_raw_cornflow_integration.py
@@ -33,32 +33,41 @@ def _get_file(relative_path):
 
 
 # Testing suit classes
-class TestCornflowClientUser(TestCase):
+class TestRawCornflowClientUser(TestCase):
     def setUp(self):
         self.client = CornFlow(url="http://127.0.0.1:5050/")
-        login_result = self.client.login("user", "UserPassword1!")
-        self.assertIn("id", login_result.keys())
-        self.assertIn("token", login_result.keys())
-        self.user_id = login_result["id"]
+        login_result = self.client.raw.login("user", "UserPassword1!")
+        data = login_result.json()
+        self.assertEqual(login_result.status_code, 200)
+        self.assertIn("id", data.keys())
+        self.assertIn("token", data.keys())
+        self.user_id = data["id"]
 
     def tearDown(self):
         pass
 
     def test_health_endpoint(self):
-        response = self.client.is_alive()
-        self.assertEqual(response["cornflow_status"], "healthy")
-        self.assertEqual(response["airflow_status"], "healthy")
+        response = self.client.raw.is_alive()
+        self.assertEqual(response.status_code, 200)
+        data = response.json()
+        self.assertEqual(data["cornflow_status"], "healthy")
+        self.assertEqual(data["airflow_status"], "healthy")
 
     def test_sign_up(self):
-        response = self.client.sign_up(
-            "test_username", "test_username@cornflow.org", "TestPassword2!"
+        response = self.client.raw.sign_up(
+            "test_username_2", "test_username_2@cornflow.org", "TestPassword2!"
         )
-        self.assertIn("id", response.keys())
-        self.assertIn("token", response.keys())
+        data = response.json()
+        self.assertIn("id", data.keys())
+        self.assertIn("token", data.keys())
+        self.assertEqual(201, response.status_code)
 
     def test_create_instance(self):
         data = _load_file(PULP_EXAMPLE)
-        response = self.client.create_instance(data, "test_example", "test_description")
+        response = self.client.raw.create_instance(data, "test_example", "test_description")
+        self.assertEqual(response.status_code, 201)
+        instance = response.json()
+
         items = [
             "id",
             "name",
@@ -70,22 +79,23 @@ class TestCornflowClientUser(TestCase):
             "executions",
         ]
         for item in items:
-            self.assertIn(item, response.keys())
+            self.assertIn(item, instance.keys())
+        self.assertEqual("test_example", instance["name"])
+        self.assertEqual("solve_model_dag", instance["schema"])
+        self.assertEqual("test_description", instance["description"])
 
-        self.assertEqual("test_example", response["name"])
-        self.assertEqual("solve_model_dag", response["schema"])
-        self.assertEqual("test_description", response["description"])
-
-        return response
+        return instance
 
     def test_create_case(self):
         data = _load_file(PULP_EXAMPLE)
-        response = self.client.create_case(
+        response = self.client.raw.create_case(
             name="test_case",
             schema="solve_model_dag",
             data=data,
             description="test_description",
         )
+        self.assertEqual(response.status_code, 201)
+        case = response.json()
 
         items = [
             "id",
@@ -102,18 +112,20 @@ class TestCornflowClientUser(TestCase):
         ]
 
         for item in items:
-            self.assertIn(item, response.keys())
-        self.assertEqual("test_case", response["name"])
-        self.assertEqual("solve_model_dag", response["schema"])
-        self.assertEqual("test_description", response["description"])
-        return response
+            self.assertIn(item, case.keys())
+        self.assertEqual("test_case", case["name"])
+        self.assertEqual("solve_model_dag", case["schema"])
+        self.assertEqual("test_description", case["description"])
+        return case
 
     def test_create_instance_file(self):
-        response = self.client.create_instance_file(
+        response = self.client.raw.create_instance_file(
             _get_file("../data/test_mps.mps"),
             name="test_filename",
             description="filename_description",
         )
+        instance = response.json()
+        self.assertEqual(response.status_code, 201)
 
         items = [
             "id",
@@ -127,21 +139,22 @@ class TestCornflowClientUser(TestCase):
         ]
 
         for item in items:
-            self.assertIn(item, response.keys())
-
-        self.assertEqual("test_filename", response["name"])
-        self.assertEqual("solve_model_dag", response["schema"])
-        self.assertEqual("filename_description", response["description"])
+            self.assertIn(item, instance.keys())
+        self.assertEqual("test_filename", instance["name"])
+        self.assertEqual("solve_model_dag", instance["schema"])
+        self.assertEqual("filename_description", instance["description"])
 
     def test_create_execution(self):
         instance = self.test_create_instance()
-        response = self.client.create_execution(
+        response = self.client.raw.create_execution(
             instance_id=instance["id"],
             config={"solver": "PULP_CBC_CMD", "timeLimit": 60},
             name="test_execution",
             description="execution_description",
             schema="solve_model_dag",
         )
+        self.assertEqual(response.status_code, 201)
+        execution = response.json()
         items = [
             "id",
             "name",
@@ -157,58 +170,65 @@ class TestCornflowClientUser(TestCase):
         ]
 
         for item in items:
-            self.assertIn(item, response.keys())
-
-        self.assertEqual(instance["id"], response["instance_id"])
-        self.assertEqual("test_execution", response["name"])
-        self.assertEqual("execution_description", response["description"])
+            self.assertIn(item, execution.keys())
+        self.assertEqual(instance["id"], execution["instance_id"])
+        self.assertEqual("test_execution", execution["name"])
+        self.assertEqual("execution_description", execution["description"])
         self.assertEqual(
-            {"solver": "PULP_CBC_CMD", "timeLimit": 60}, response["config"]
+            {"solver": "PULP_CBC_CMD", "timeLimit": 60}, execution["config"]
         )
-        self.assertEqual(STATUS_QUEUED, response["state"])
+        self.assertEqual(STATUS_QUEUED, execution["state"])
 
-        return response
+        return execution
 
     def test_create_execution_data_check(self):
         exec_to_check = self.test_create_execution()
         time.sleep(15)
         exec_to_check_id = exec_to_check["id"]
-        execution = self.client.create_execution_data_check(exec_to_check_id)
+        response = self.client.raw.create_execution_data_check(exec_to_check_id)
+        execution = response.json()
         self.assertEqual(STATUS_QUEUED, execution["state"])
         return execution
 
     def test_execution_data_check_solution(self):
         execution = self.test_create_execution_data_check()
         time.sleep(15)
-        results = self.client.get_solution(execution["id"])
-        self.assertEqual(results["state"], 1)
+        results = self.client.raw.get_solution(execution["id"])
+        self.assertEqual(results.status_code, 200)
+        self.assertEqual(results.json()["state"], 1)
 
     def test_create_instance_data_check(self):
         inst_to_check = self.test_create_instance()
         inst_to_check_id = inst_to_check["id"]
-        execution = self.client.create_instance_data_check(inst_to_check_id)
-        self.assertEqual(STATUS_QUEUED, execution["state"])
-        config = execution.get("config")
+        response = self.client.raw.create_instance_data_check(inst_to_check_id)
+        self.assertEqual(response.status_code, 201)
+        data = response.json()
+        self.assertEqual(STATUS_QUEUED, data["state"])
+        config = data.get("config")
         self.assertIsInstance(config, dict)
         self.assertTrue(config.get("checks_only"))
-        self.assertEqual(execution.get("schema"), "solve_model_dag")
-        self.assertEqual(execution.get("instance_id"), inst_to_check_id)
-        return execution
+        self.assertEqual(data.get("schema"), "solve_model_dag")
+        self.assertEqual(data.get("instance_id"), inst_to_check_id)
+        return data
 
     def test_instance_data_check_solution(self):
         execution = self.test_create_instance_data_check()
         time.sleep(15)
-        results = self.client.get_solution(execution["id"])
-        self.assertEqual(results["state"], 1)
-        response = self.client.get_one_instance_data(
-            reference_id=execution["instance_id"], encoding="br"
+        results = self.client.raw.get_solution(execution["id"])
+        self.assertEqual(results.status_code, 200)
+        self.assertEqual(results.json()["state"], 1)
+        response = self.client.raw.get_api_for_id(
+            api="instance", id=execution["instance_id"], post_url="data", encoding="br"
         )
-        self.assertIsNotNone(response["checks"])
+        self.assertEqual(response.status_code, 200)
+        self.assertIsNotNone(response.json()["checks"])
 
     def test_execution_results(self):
         execution = self.test_create_execution()
         time.sleep(10)
-        response = self.client.get_results(execution["id"])
+        response = self.client.raw.get_results(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
 
         items = [
             "id",
@@ -235,27 +255,32 @@ class TestCornflowClientUser(TestCase):
         self.assertEqual(STATUS_QUEUED, execution["state"])
 
         time.sleep(4)
-        response = self.client.get_status(execution["id"])
+        response = self.client.raw.get_status(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = ["id", "state", "message", "data_hash"]
         for item in items:
             self.assertIn(item, response.keys())
         self.assertEqual(STATUS_NOT_SOLVED, response["state"])
 
         time.sleep(10)
-        response = self.client.get_status(execution["id"])
+        response = self.client.raw.get_status(execution["id"])
+        self.assertEqual(response.status_code, 200)
         for item in items:
-            self.assertIn(item, response.keys())
-        self.assertEqual(STATUS_OPTIMAL, response["state"])
+            self.assertIn(item, response.json().keys())
+        self.assertEqual(STATUS_OPTIMAL, response.json()["state"])
 
     def test_stop_execution(self):
         execution = self.test_create_execution()
-        response = self.client.stop_execution(execution["id"])
-        self.assertEqual(response["message"], "The execution has been stopped")
+        response = self.client.raw.stop_execution(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json()["message"], "The execution has been stopped")
 
     def test_get_execution_log(self):
         execution = self.test_create_execution()
-        response = self.client.get_log(execution["id"])
-
+        response = self.client.raw.get_log(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -278,7 +303,9 @@ class TestCornflowClientUser(TestCase):
     def test_get_execution_solution(self):
         execution = self.test_create_execution()
         time.sleep(15)
-        response = self.client.get_solution(execution["id"])
+        response = self.client.raw.get_solution(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -305,22 +332,26 @@ class TestCornflowClientUser(TestCase):
 
     def test_put_one_execution(self):
         execution = self.test_create_execution()
-        response = self.client.put_one_execution(execution["id"], {"name": "new_execution_name"})
-        self.assertEqual("Updated correctly", response["message"])
+        response = self.client.raw.put_one_execution(execution["id"], {"name": "new_execution_name"})
+        self.assertEqual(200, response.status_code)
+        self.assertEqual("Updated correctly", response.json()["message"])
 
     def test_delete_one_execution(self):
         execution = self.test_create_execution()
-        response = self.client.delete_one_execution(execution["id"])
-        self.assertEqual("The object has been deleted", response["message"])
+        response = self.client.raw.delete_one_execution(execution["id"])
+        self.assertEqual(200, response.status_code)
+        self.assertEqual("The object has been deleted", response.json()["message"])
 
     def test_create_case_execution(self):
         execution = self.test_get_execution_solution()
-        response = self.client.create_case(
+        response = self.client.raw.create_case(
             name="case_from_solution",
             schema="solve_model_dag",
             description="case_from_solution_description",
             solution=execution["data"],
         )
+        self.assertEqual(response.status_code, 201)
+        response = response.json()
 
         items = [
             "id",
@@ -347,24 +378,28 @@ class TestCornflowClientUser(TestCase):
     def test_get_all_instances(self):
         self.test_create_instance()
         self.test_create_instance()
-        instances = self.client.get_all_instances()
-        self.assertGreaterEqual(len(instances), 2)
+        instances = self.client.raw.get_all_instances()
+        self.assertEqual(instances.status_code, 200)
+        self.assertGreaterEqual(len(instances.json()), 2)
 
     def test_get_all_executions(self):
         self.test_stop_execution()
         self.test_stop_execution()
-        executions = self.client.get_all_executions()
-        self.assertGreaterEqual(len(executions), 2)
+        executions = self.client.raw.get_all_executions()
+        self.assertEqual(executions.status_code, 200)
+        self.assertGreaterEqual(len(executions.json()), 2)
 
     def test_get_all_cases(self):
         self.test_create_case()
         self.test_create_case()
-        cases = self.client.get_all_cases()
-        self.assertGreaterEqual(len(cases), 2)
+        cases = self.client.raw.get_all_cases()
+        self.assertEqual(cases.status_code, 200)
+        self.assertGreaterEqual(len(cases.json()), 2)
 
     def test_get_one_user(self):
-        response = self.client.get_one_user(self.user_id)
-
+        response = self.client.raw.get_one_user(self.user_id)
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = ["id", "first_name", "last_name", "username", "email"]
         for item in items:
             self.assertIn(item, response.keys())
@@ -375,7 +410,9 @@ class TestCornflowClientUser(TestCase):
 
     def test_get_one_instance(self):
         instance = self.test_create_instance()
-        response = self.client.get_one_instance(instance["id"])
+        response = self.client.raw.get_one_instance(instance["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -390,10 +427,12 @@ class TestCornflowClientUser(TestCase):
         for item in items:
             self.assertIn(item, response.keys())
             self.assertEqual(instance[item], response[item])
-            
+
     def test_get_one_instance_data(self):
         instance = self.test_create_instance()
-        response = self.client.get_one_instance_data(instance["id"])
+        response = self.client.raw.get_one_instance_data(instance["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -411,12 +450,15 @@ class TestCornflowClientUser(TestCase):
 
     def test_put_one_instance(self):
         instance = self.test_create_instance()
-        response = self.client.put_one_instance(instance["id"], {"name": "new_instance_name"})
-        self.assertEqual("Updated correctly", response["message"])
+        response = self.client.raw.put_one_instance(instance["id"], {"name": "new_instance_name"})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual("Updated correctly", response.json()["message"])
 
     def test_get_one_case(self):
         case = self.test_create_case()
-        response = self.client.get_one_case(case["id"])
+        response = self.client.raw.get_one_case(case["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -434,10 +476,12 @@ class TestCornflowClientUser(TestCase):
         for item in items:
             self.assertIn(item, response.keys())
             self.assertEqual(case[item], response[item])
-            
+
     def test_get_one_case_data(self):
         case = self.test_create_case()
-        response = self.client.get_one_case_data(case["id"])
+        response = self.client.raw.get_one_case_data(case["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -459,7 +503,9 @@ class TestCornflowClientUser(TestCase):
 
     def test_get_one_case_execution(self):
         case = self.test_create_case_execution()
-        response = self.client.get_one_case(case["id"])
+        response = self.client.raw.get_one_case(case["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = [
             "id",
             "name",
@@ -480,13 +526,15 @@ class TestCornflowClientUser(TestCase):
 
     def test_delete_one_case(self):
         case = self.test_create_case()
-        response = self.client.delete_one_case(case["id"])
-        self.assertEqual("The object has been deleted", response["message"])
+        response = self.client.raw.delete_one_case(case["id"])
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual("The object has been deleted", response.json()["message"])
 
     def test_put_one_case(self):
         case = self.test_create_case()
-        response = self.client.put_one_case(case["id"], {"name": "new_case_name"})
-        self.assertEqual("Updated correctly", response["message"])
+        response = self.client.raw.put_one_case(case["id"], {"name": "new_case_name"})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual("Updated correctly", response.json()["message"])
 
     def test_patch_one_case(self):
         # TODO: Get example of data patch for the tests
@@ -494,11 +542,14 @@ class TestCornflowClientUser(TestCase):
 
     def test_delete_one_instance(self):
         instance = self.test_create_instance()
-        response = self.client.delete_one_instance(instance["id"])
-        self.assertEqual("The object has been deleted", response["message"])
+        response = self.client.raw.delete_one_instance(instance["id"])
+        self.assertEqual(200, response.status_code)
+        self.assertEqual("The object has been deleted", response.json()["message"])
 
     def test_get_schema(self):
-        response = self.client.get_schema("solve_model_dag")
+        response = self.client.raw.get_schema("solve_model_dag")
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         schema = {
             "config": get_pulp_jsonschema("solver_config.json"),
             "instance": get_pulp_jsonschema(),
@@ -512,14 +563,14 @@ class TestCornflowClientUser(TestCase):
             self.assertDictEqual(value, response[key])
 
     def test_get_all_schemas(self):
-        response = self.client.get_all_schemas()
-        read_schemas = [dag for v in response for (_, dag) in v.items()]
+        response = self.client.raw.get_all_schemas()
+        read_schemas = [dag for v in response.json() for (_, dag) in v.items()]
 
         for schema in PUBLIC_DAGS:
             self.assertIn(schema, read_schemas)
 
 
-class TestCornflowClientAdmin(TestCase):
+class TestRawCornflowClientAdmin(TestCase):
     def setUp(self):
         self.client = CornFlow(url="http://127.0.0.1:5050/")
         login_result = self.client.login("admin", "Adminpassword1!")
@@ -533,22 +584,23 @@ class TestCornflowClientAdmin(TestCase):
         pass
 
     def test_get_all_users(self):
-        response = self.client.get_all_users()
-        self.assertGreaterEqual(len(response), 3)
+        response = self.client.raw.get_all_users()
+        self.assertGreaterEqual(len(response.json()), 3)
 
     def test_get_one_user(self):
-        response = self.client.get_one_user(self.base_user_id)
+        response = self.client.raw.get_one_user(self.base_user_id)
+        self.assertEqual(response.status_code, 200)
 
         items = ["id", "first_name", "last_name", "username", "email"]
         for item in items:
-            self.assertIn(item, response.keys())
+            self.assertIn(item, response.json().keys())
 
-        self.assertEqual(self.base_user_id, response["id"])
-        self.assertEqual("user", response["username"])
-        self.assertEqual("user@cornflow.org", response["email"])
+        self.assertEqual(self.base_user_id, response.json()["id"])
+        self.assertEqual("user", response.json()["username"])
+        self.assertEqual("user@cornflow.org", response.json()["email"])
 
 
-class TestCornflowClientService(TestCase):
+class TestRawCornflowClientService(TestCase):
     def setUp(self):
         self.client = CornFlow(url="http://127.0.0.1:5050/")
         login_result = self.client.login("airflow", "Airflow_test_password1")
@@ -562,15 +614,17 @@ class TestCornflowClientService(TestCase):
         client = CornFlow(url="http://127.0.0.1:5050/")
         _ = client.login("user", "UserPassword1!")
         data = _load_file(PULP_EXAMPLE)
-        instance = client.create_instance(data, "test_example", "test_description")
-        execution = client.create_execution(
+        instance = client.raw.create_instance(data, "test_example", "test_description").json()
+        execution = client.raw.create_execution(
             instance_id=instance["id"],
             config={"solver": "PULP_CBC_CMD", "timeLimit": 60},
             name="test_execution",
             description="execution_description",
             schema="solve_model_dag",
-        )
-        response = self.client.get_data(execution["id"])
+        ).json()
+        response = self.client.raw.get_data(execution["id"])
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
         items = ["id", "data", "config"]
 
         for item in items:
@@ -584,29 +638,31 @@ class TestCornflowClientService(TestCase):
         client = CornFlow(url="http://127.0.0.1:5050/")
         _ = client.login("user", "UserPassword1!")
         data = _load_file(PULP_EXAMPLE)
-        instance = client.create_instance(data, "test_example", "test_description")
-        execution = client.create_execution(
+        instance = client.raw.create_instance(data, "test_example", "test_description").json()
+        execution = client.raw.create_execution(
             instance_id=instance["id"],
             config={"solver": "PULP_CBC_CMD", "timeLimit": 60},
             name="test_execution",
             description="execution_description",
             schema="solve_model_dag",
-        )
+        ).json()
 
         time.sleep(15)
 
-        solution = client.get_solution(execution["id"])
+        solution = client.raw.get_solution(execution["id"]).json()
 
         payload = dict(
             state=1, log_json={}, log_text="", solution_schema="solve_model_dag"
         )
         payload["data"] = solution["data"]
 
-        response = self.client.write_solution(execution["id"], **payload)
-        self.assertEqual("results successfully saved", response["message"])
+        response = self.client.raw.write_solution(execution["id"], **payload)
+        self.assertEqual("results successfully saved", response.json()["message"])
 
     def test_get_deployed_dags(self):
-        response = self.client.get_deployed_dags()
+        response = self.client.raw.get_deployed_dags()
+        self.assertEqual(response.status_code, 200)
+        response = response.json()
 
         items = ["id", "description"]
         for item in items:
@@ -619,13 +675,15 @@ class TestCornflowClientService(TestCase):
 
     def test_post_deployed_dag(self):
 
-        response = self.client.create_deployed_dag(
-            name="test_dag", description="test_dag_description"
+        response = self.client.raw.create_deployed_dag(
+            name="test_dag_2", description="test_dag_2_description"
         )
+        self.assertEqual(response.status_code, 201)
+        response = response.json()
 
         items = ["id", "description"]
 
         for item in items:
             self.assertIn(item, response.keys())
-        self.assertEqual("test_dag", response["id"])
-        self.assertEqual("test_dag_description", response["description"])
+        self.assertEqual("test_dag_2", response["id"])
+        self.assertEqual("test_dag_2_description", response["description"])
