+++ libs/core/setup.py
@@ -9,7 +9,7 @@ with open("requirements.txt", "r") as fh:
 
 setuptools.setup(
     name="cornflow-core",
-    version="0.1.3",
+    version="0.1.4",
     author="baobab soluciones",
     author_email="sistemas@baobabsoluciones.es",
     description="REST API flask backend components used by cornflow and other REST APIs",
